Content

Content

Content

Details

An R6 class that represents content

See also

Public fields

connect

An R6 Connect object

content

The content details from Posit Connect

Active bindings

default_variant

The default variant for this object.

is_rendered

TRUE if this is a rendered content type, otherwise FALSE.

is_interactive

TRUE if this is a rendered content type, otherwise FALSE.

Methods

Public methods

Content$new()

Content$get_connect()

Content$get_content()

Content$get_content_remote()

Content$get_bundles()

Content$bundle_download()

Content$bundle_delete()

Content$internal_content()

Content$update()

Content$danger_delete()

Content$get_url()

Content$get_dashboard_url()

Content$jobs()

Content$job()

Content$variants()

Content$tag_set()

Content$tag_delete()

Content$tags()

Content$permissions_add()

Content$permissions_update()

Content$permissions_delete()

Content$permissions()

Content$environment()

Content$environment_set()

Content$environment_all()

Content$deploy()

Content$repo_enable()

Content$repo_set()

Content$print()

Content$clone()

Method new()

Initialize this content.

Usage

Content$new(connect, content)

Arguments

connect

The Connect instance.

content

The content data.

Method get_connect()

Returns the Connect instance.

Usage

Content$get_connect()

Method get_content()

Returns the underlying content data.

Usage

Content$get_content()

Method get_content_remote()

Obtain the content data from the Connect server.

Usage

Content$get_content_remote()

Method get_bundles()

Return the set of content bundles.

Usage

Content$get_bundles()

Method bundle_download()

Download the source archive for a content bundle.

Usage

Content$bundle_download(
  bundle_id,
  filename = tempfile(pattern = "bundle", fileext = ".tar.gz"),
  overwrite = FALSE
)

Arguments

bundle_id

The bundle identifer.

filename

Where to write the result.

overwrite

Overwrite an existing filename.

Method bundle_delete()

Delete a content bundle.

Usage

Content$bundle_delete(bundle_id)

Arguments

bundle_id

The bundle identifer.

Method internal_content()

Get this (remote) content item.

Usage

Content$internal_content()

Method update()

Update this content item.

Usage

Content$update(...)

Arguments

...

Content fields.

Method danger_delete()

Delete this content item.

Usage

Content$danger_delete()

Method get_url()

Return the URL for this content.

Usage

Content$get_url()

Method get_dashboard_url()

Return the URL for this content in the Posit Connect dashboard.

Usage

Content$get_dashboard_url(pane = "")

Arguments

pane

The pane in the dashboard to link to.

Method jobs()

Return the jobs for this content.

Usage

Content$jobs()

Method job()

Return a single job for this content.

Usage

Content$job(key)

Arguments

key

The job key.

Method variants()

Return the variants for this content.

Usage

Content$variants()

Method tag_set()

Set a tag for this content.

Usage

Content$tag_set(tag_id)

Arguments

tag_id

The tag identifier.

Method tag_delete()

Remove a tag for this content.

Usage

Content$tag_delete(tag_id)

Arguments

tag_id

The tag identifier.

Method tags()

The tags for this content.

Usage

Content$tags()

Method permissions_add()

Add a principal to the ACL for this content.

Usage

Content$permissions_add(principal_guid, principal_type, role)

Arguments

principal_guid

GUID for the target user or group.

principal_type

Acting on user or group.

role

The kind of content access.

Method permissions_update()

Alter a principal in the ACL for this content.

Usage

Content$permissions_update(id, principal_guid, principal_type, role)

Arguments

id

The target identifier.

principal_guid

GUID for the target user or group.

principal_type

Acting on user or group.

role

The kind of content access.

Method permissions_delete()

Remove an entry from the ACL for this content.

Usage

Content$permissions_delete(id)

Arguments

id

The target identifier.

Method permissions()

Obtain some or all of the ACL for this content.

Usage

Content$permissions(id = NULL, add_owner = FALSE)

Arguments

id

The target identifier.

add_owner

Include the content owner in the result set.

Method environment()

Return the environment variables set for this content.

Usage

Content$environment()

Method environment_set()

Adjust the environment variables set for this content.

Usage

Content$environment_set(...)

Arguments

...

Environment variable names and values. Use NA as the value to unset variables.

Method environment_all()

Overwrite the environment variables set for this content.

Usage

Content$environment_all(...)

Arguments

...

Environment variable names and values.

Method deploy()

Deploy this content

Usage

Content$deploy(bundle_id = NULL)

Arguments

bundle_id

Target bundle identifier.

Method repo_enable()

Adjust Git polling.

Usage

Content$repo_enable(enabled = TRUE)

Arguments

enabled

Polling enabled.

Method repo_set()

Adjust Git repository.

Usage

Content$repo_set(repository, branch, subdirectory)

Arguments

repository

Git repository URL

branch

Git repository branch

subdirectory

Git repository directory

Method print()

Print this object.

Usage

Content$print(...)

Arguments

...

Unused.

Method clone()

The objects of this class are cloneable with this method.

Usage

Content$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.