Class representing a Connect API client

Class representing a Connect API client

Class representing a Connect API client

Usage

client <- Connect$new(host = 'connect.example.com',
  apiKey = 'mysecretkey')
client$get_apps()
client$get_tags()

Details

This class allows a user to interact with a Connect server via the Connect API. Authentication is done by providing an API key.

See also

Methods

Public methods

Connect$get_connect()

Connect$new()

Connect$print()

Connect$raise_error()

Connect$add_auth()

Connect$GET()

Connect$GET_RESULT()

Connect$GET_URL()

Connect$GET_RESULT_URL()

Connect$PUT()

Connect$HEAD()

Connect$DELETE()

Connect$POST()

Connect$me()

Connect$get_tags()

Connect$get_tag_id()

Connect$get_tag_tree()

Connect$create_tag()

Connect$get_n_apps()

Connect$get_apps()

Connect$get_schedule()

Connect$content_create()

Connect$download_bundle()

Connect$content_upload()

Connect$content_deploy()

Connect$content()

Connect$task()

Connect$set_content_tag()

Connect$user()

Connect$users()

Connect$users_remote()

Connect$users_create()

Connect$users_lock()

Connect$users_unlock()

Connect$users_update()

Connect$groups()

Connect$group_members()

Connect$group_member_add()

Connect$group_member_remove()

Connect$groups_create()

Connect$inst_content_visits()

Connect$inst_shiny_usage()

Connect$procs()

Connect$docs()

Connect$audit_logs()

Connect$server_settings_r()

Connect$server_settings()

Connect$clone()

Method get_connect()

Usage

Connect$get_connect()

Method new()

Usage

Connect$new(host, api_key)

Method print()

Usage

Connect$print(...)

Method raise_error()

Usage

Connect$raise_error(res)

Method add_auth()

Usage

Connect$add_auth()

Method GET()

Usage

Connect$GET(path, writer = httr::write_memory(), parser = "parsed")

Method GET_RESULT()

Usage

Connect$GET_RESULT(path, writer = httr::write_memory())

Method GET_URL()

Usage

Connect$GET_URL(url, writer = httr::write_memory(), parser = "parsed")

Method GET_RESULT_URL()

Usage

Connect$GET_RESULT_URL(url, writer = httr::write_memory())

Method PUT()

Usage

Connect$PUT(path, body, encode = "json")

Method HEAD()

Usage

Connect$HEAD(path)

Method DELETE()

Usage

Connect$DELETE(path)

Method POST()

Usage

Connect$POST(path, body, encode = "json", prefix = "/__api__/")

Method me()

Usage

Connect$me()

Method get_tags()

Usage

Connect$get_tags(use_cache = FALSE)

Method get_tag_id()

Usage

Connect$get_tag_id(tagname)

Method get_tag_tree()

Usage

Connect$get_tag_tree()

Method create_tag()

Usage

Connect$create_tag(name, parent_id = NULL)

Method get_n_apps()

Usage

Connect$get_n_apps()

Method get_apps()

Usage

Connect$get_apps(filter = NULL, .collapse = "&", .limit = Inf, page_size = 25)

Method get_schedule()

Usage

Connect$get_schedule(schedule_id)

Method content_create()

Usage

Connect$content_create(name, title = name, ...)

Method download_bundle()

Usage

Connect$download_bundle(bundle_id, to_path = tempfile())

Method content_upload()

Usage

Connect$content_upload(bundle_path, guid)

Method content_deploy()

Usage

Connect$content_deploy(guid, bundle_id)

Method content()

Usage

Connect$content(guid)

Method task()

Usage

Connect$task(task_id, first = 0, wait = 5)

Method set_content_tag()

Usage

Connect$set_content_tag(content_id, tag_id)

Method user()

Usage

Connect$user(guid)

Method users()

Usage

Connect$users(page_number = 1, prefix = NULL, page_size = 20)

Method users_remote()

Usage

Connect$users_remote(prefix)

Method users_create()

Usage

Connect$users_create(
  username,
  email,
  first_name = NULL,
  last_name = NULL,
  password = NULL,
  user_must_set_password = NULL,
  user_role = NULL
)

Method users_lock()

Usage

Connect$users_lock(user_guid)

Method users_unlock()

Usage

Connect$users_unlock(user_guid)

Method users_update()

Usage

Connect$users_update(user_guid, ...)

Method groups()

Usage

Connect$groups(page_number = 1, prefix = NULL, page_size = 20)

Method group_members()

Usage

Connect$group_members(guid)

Method group_member_add()

Usage

Connect$group_member_add(group_guid, user_guid)

Method group_member_remove()

Usage

Connect$group_member_remove(group_guid, user_guid)

Method groups_create()

Usage

Connect$groups_create(name)

Method inst_content_visits()

Usage

Connect$inst_content_visits(
  content_guid = NULL,
  min_data_version = NULL,
  from = NULL,
  to = NULL,
  limit = 20,
  previous = NULL,
  nxt = NULL,
  asc_order = TRUE
)

Method inst_shiny_usage()

Usage

Connect$inst_shiny_usage(
  content_guid = NULL,
  min_data_version = NULL,
  from = NULL,
  to = NULL,
  limit = 20,
  previous = NULL,
  nxt = NULL,
  asc_order = TRUE
)

Method procs()

Usage

Connect$procs()

Method docs()

Usage

Connect$docs(docs = "api", browse = TRUE)

Method audit_logs()

Usage

Connect$audit_logs(limit = 20L, previous = NULL, nxt = NULL, asc_order = TRUE)

Method server_settings_r()

Usage

Connect$server_settings_r()

Method server_settings()

Usage

Connect$server_settings()

Method clone()

The objects of this class are cloneable with this method.

Usage

Connect$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.