Update Content

Update settings for a content item. For a list of all settings, see the latest documentation or the documentation for your server via connectapi::browse_api_docs().

Usage

content_update(content, ...)

content_update_access_type(content, access_type = c("all", "logged_in", "acl"))

content_update_owner(content, owner_guid)

Arguments

content

An R6 content item

...

Settings up update that are passed along to Posit Connect

access_type

One of "all", "logged_in", or "acl"

owner_guid

The GUID of a user who is a publisher, so that they can become the new owner of the content

Value

An R6 content item

Details

Popular selections are content_update(access_type="all"), content_update(access_type="logged_in") or content_update(access_type="acl"), process settings, title, description, etc.

content_update_access_type() is a helper to make it easier to change access_type

content_update_owner() is a helper to make it easier to change owner

See also