Deploy a Git Repository

Deploy a git repository directly to Posit Connect, using Posit Connect's "pull-based" "git-polling" method of deployment.

Usage

deploy_repo(
  client,
  repository,
  branch,
  subdirectory,
  name = create_random_name(),
  title = name,
  ...
)

deploy_repo_enable(content, enabled = TRUE)

deploy_repo_update(content)

Arguments

client

A Connect R6 object

repository

The git repository to deploy

branch

The git branch to deploy

subdirectory

The subdirectory to deploy (must contain a manifest.json)

name

The "name" / unique identifier for the content. Defaults to a random character string

title

The "title" of the content

...

Additional options for defining / specifying content attributes

content

An R6 Content object (i.e. the result of content_item())

enabled

Whether Connect will enable automatic polling for repository updates

Value

A ContentTask object, for use with poll_task() (if you want to follow the logs)

Details

deploy_repo_enable() enables (or disables) Posit Connect's git polling for a piece of content

deploy_repo_update() triggers an update of the content from its git repository, if any are present

See also