Get ACL Details

Retrieve the Access Controls associated with a given piece of content.

get_acl_user(content)

get_acl_group(content)

get_acl(content)

get_acl_user_role(content, user_guid)

get_acl_group_role(content, group_guid)

Arguments

content
Content An R6 Content item as returned from content_item()

user_guid
character A user guid. Get user guids using get_users()

group_guid
character A group guid. Get group guids using get_groups()

Value

A list of users/groups who have access to the content

Details

NOTE: ACLs can still be stored, even when access_type for content is "all" or "logged_in" users. In these cases, granting or removing "viewer" privileges have no effect.

get_acl_user() returns user ACLs

get_acl_group() returns group ACLs

get_acl_user_role() returns the "role" for a particular user on a piece of content

get_acl_group_role() returns the "role" for a particular group on a piece of content

get_acl() is deprecated.

See also