Get information about content on the RStudio Connect server

Get information about content on the RStudio Connect server

get_content(src, filter = NULL, limit = 25, page_size = 25)

Arguments

src
The source object

filter
a named list of filter options, e.g. list(name = 'appname')

limit
the maximum number of records to return

page_size
the number of records to return per page

Value

A tibble with the following columns:

idThe application ID

guidThe unique identifier of this content item.

access_typeAccess type describes how this content manages its viewers. The value all is the most permissive; any visitor to RStudio Connect will be able to view this content. The value logged_in indicates that all RStudio Connect accounts may view the content. The acl value lets specifically enumerated users and groups view the content. Users configured as collaborators may always view content. It may have a value of all, logged_in or acl.

connection_timeoutMaximum number of seconds allowed without data sent or received across a client connection. A value of 0 means connections will never time-out (not recommended). When null, the default Scheduler.ConnectionTimeout is used. Applies only to content types that are executed on demand.

read_timeoutMaximum number of seconds allowed without data received from a client connection. A value of 0 means a lack of client (browser) interaction never causes the connection to close. When null, the default Scheduler.ReadTimeout is used. Applies only to content types that are executed on demand.

init_timeoutThe maximum number of seconds allowed for an interactive application to start. RStudio Connect must be able to connect to a newly launched Shiny application, for example, before this threshold has elapsed. When null, the default Scheduler.InitTimeout is used. Applies only to content types that are executed on demand.

idle_timeoutThe maximum number of seconds a worker process for an interactive application to remain alive after it goes idle (no active connections). When null, the default Scheduler.IdleTimeout is used. Applies only to content types that are executed on demand.

max_processesSpecifies the total number of concurrent processes allowed for a single interactive application. When null, the default Scheduler.MaxProcesses is used. Applies only to content types that are executed on demand.

min_processesSpecifies the minimum number of concurrent processes allowed for a single interactive application. When null, the default Scheduler.MinProcesses is used. Applies only to content types that are executed on demand.

max_conns_per_processSpecifies the maximum number of client connections allowed to an individual process. Incoming connections which will exceed this limit are routed to a new process or rejected. When null, the default Scheduler.MaxConnsPerProcess is used. Applies only to content types that are executed on demand.

load_factorControls how aggressively new processes are spawned. When null, the default Scheduler.LoadFactor is used. Applies only to content types that are executed on demand.

urlThe URL associated with this content. Computed from the associated vanity URL or the identifiers for this content.

vanity_urlThe vanity url assigned to this app by an administrator

nameA simple, URL-friendly identifier. Allows alpha-numeric characters, hyphens ("-"), and underscores ("_").

titleThe title of this content.

bundle_idThe identifier for the active deployment bundle. Automatically assigned upon the successful deployment of that bundle.

app_modeThe runtime model for this content. Has a value of unknown before data is deployed to this item. Automatically assigned upon the first successful bundle deployment.

content_categoryDescribes the specialization of the content runtime model. Automatically assigned upon the first successful bundle deployment.

has_parametersTrue when R Markdown rendered content allows parameter configuration. Automatically assigned upon the first successful bundle deployment. Applies only to content with an app_mode of rmd-static.

created_timeThe timestamp (RFC3339) indicating when this content was created.

last_deployed_timeThe timestamp (RFC3339) indicating when this content last had a successful bundle deployment performed.

r_versionThe version of the R interpreter associated with this content. The value null represents that an R interpreter is not used by this content or that the R package environment has not been successfully restored. Automatically assigned upon the successful deployment of a bundle.

py_versionThe version of the Python interpreter associated with this content. The value null represents that a Python interpreter is not used by this content or that the Python package environment has not been successfully restored. Automatically assigned upon the successful deployment of a bundle.

build_status

run_asThe UNIX user that executes this content. When null, the default Applications.RunAs is used. Applies only to executable content types - not static.

run_as_current_userIndicates if this content is allowed to execute as the logged-in user when using PAM authentication. Applies only to executable content types - not static.

descriptionA rich description of this content

app_roleThe relationship of the accessing user to this content. A value of owner is returned for the content owner. editor indicates a collaborator. The viewer value is given to users who are permitted to view the content. A none role is returned for administrators who cannot view the content but are permitted to view its configuration. Computed at the time of the request.

owner_first_nameThe first name of the owner of the content.

owner_last_nameThe last name of the owner of the content.

owner_usernameThe username of the owner of the content.

owner_guidThe unique identifier for the owner

owner_emailThe email of the content owner

owner_lockedIs the owners user account locked?

is_scheduledIs this content scheduled?

gitIs this content deployed via git or GitHub?

Details

Please see https://docs.rstudio.com/connect/api/#getContent for more information

Examples

if (FALSE) {
  library(connectapi)
  client <- connect()

  get_content(client, limit = 20)
}