Get users within a specific group

Get users within a specific group

get_group_members(src, guid)

Arguments

src
The source object

guid
A group GUID identifier

Value

A tibble with the following columns:

emailThe user's email

usernameThe user's username

first_nameThe user's first name

last_nameThe user's last name

user_roleThe user's role. It may have a value of administrator, publisher or viewer.

created_timeThe timestamp (in RFC3339 format) when the user was created in the RStudio Connect server

updated_timeThe timestamp (in RFC3339 format) when the user was last updated in the RStudio Connect server

active_timeThe timestamp (in RFC3339 format) when the user was last active on the RStudio Connect server

confirmedWhen false, the created user must confirm their account through an email. This feature is unique to password authentication.

lockedWhether or not the user is locked

guidThe user's GUID, or unique identifier, in UUID RFC4122 format

Details

Please see https://docs.rstudio.com/connect/api/#getGroupMembers for more information

Examples

if (FALSE) {
library(connectapi)
client <- connect()

# get the first 20 groups
groups <- get_groups(client)

group_guid <- groups$guid[1]

get_group_members(client, guid = group_guid)
}