Get Job Log

Get the log output for a job. Requires Connect 2022.10.0 or newer.

Usage

get_log(job, max_log_lines = NULL)

Arguments

job

A job, represented by an element from the list returned by get_job_list().

max_log_lines

Optional. An integer indicating the maximum number of log lines to return. If NULL (default), Connect returns a maximum of 5000 lines.

Value

A data frame with the requested log. Each row represents an entry.

source: stdout or stderr

timestamp: The time of the entry.

data: The logged text.

Details

Note: The output of get_jobs() cannot be used with get_log(). Please use an object from the list returned by get_job_list().

See also

Examples

if (FALSE) { # \dontrun{
client <- connect()
item <- content_item(client, "951bf3ad-82d0-4bca-bba8-9b27e35c49fa")
jobs <- get_job_list(item)
log <- get_log(jobs[[1]])
} # }