Set the Content Image

Please use set_thumbnail instead.

Set the Content Image using a variety of methods.

Usage

set_image_path(content, path)

set_image_url(content, url)

set_image_webshot(content, ...)

Arguments

content

A content object

path

The path to an image on disk

url

The url for an image

...

Additional arguments passed on to webshot2::webshot()

Details

NOTE: set_image_webshot() requires webshot2::webshot(), but currently skips and warns for any content that requires authentication until the webshot2 package supports authentication.

See also