Tag Page iframe

Build a Tag Page that shows content in iframes, so it is easy to browse between content. We also use a JQuery library to ensure content / iframes only load resources when accessed (for performance)

tag_page_iframe(connect, tag, metadata = NULL, quiet = FALSE)

Arguments

connect

A Connect object

tag

A tag to search for. NOTE: tag names are not necessarily unique

metadata

A metadata object to be used for enriching application data

quiet

Whether to suppress rmarkdown logging

Details

Unfortunately, iframes do not pass cookies as cleanly as I would like... I am having trouble even getting this to work in the "same-origin" setup that should work (it is failing if the iframe uses a different origin from the parent)

See the examples for how to define a metadata object. It should be a list of lists (i.e. a list of objects). Each "object" should have a "guid" attribute, which will be used to update the application data.

References

https://security.stackexchange.com/questions/182518/how-to-confirm-that-an-embedded-iframe-can-read-cookies-from-parent

https://www.html5rocks.com/en/tutorials/security/sandboxed-iframes/

https://stackoverflow.com/questions/31184505/sandboxing-iframe-and-allow-same-origin

https://stackoverflow.com/questions/2117248/setting-cookie-in-iframe-different-domain

https://stackoverflow.com/questions/13432821/is-it-possible-to-add-request-headers-to-an-iframe-src-request

Examples

app_metadata <- list(
  list(guid = "abcde-abcd-abcd-abcde", content_group = "Group 1"),
  list(guid = "fghij-fghi-fghi-fghij", content_group = "Group 2")
)