Variant

Variant

Variant

Details

An R6 class that represents a Variant

See also

Super class

connectapi::Content -> Variant

Public fields

key

The variant key.

variant

The variant.

Methods

Public methods

Variant$get_variant()

Variant$get_variant_remote()

Variant$new()

Variant$send_mail()

Variant$get_schedule()

Variant$get_schedule_remote()

Variant$get_subscribers()

Variant$remove_subscriber()

Variant$add_subscribers()

Variant$render()

Variant$renderings()

Variant$update_variant()

Variant$jobs()

Variant$job()

Variant$get_url()

Variant$get_url_rev()

Variant$get_dashboard_url()

Variant$print()

Variant$clone()

Inherited methods

connectapi::Content$bundle_delete()

connectapi::Content$bundle_download()

connectapi::Content$danger_delete()

connectapi::Content$deploy()

connectapi::Content$environment()

connectapi::Content$environment_all()

connectapi::Content$environment_set()

connectapi::Content$get_bundles()

connectapi::Content$get_connect()

connectapi::Content$get_content()

connectapi::Content$get_content_remote()

connectapi::Content$get_job()

connectapi::Content$get_jobs()

connectapi::Content$internal_content()

connectapi::Content$permissions()

connectapi::Content$permissions_add()

connectapi::Content$permissions_delete()

connectapi::Content$permissions_update()

connectapi::Content$repo_enable()

connectapi::Content$repo_set()

connectapi::Content$runas()

connectapi::Content$tag_delete()

connectapi::Content$tag_set()

connectapi::Content$tags()

connectapi::Content$update()

connectapi::Content$variants()

Method get_variant()

Get the underlying variant data.

Usage

Variant$get_variant()

Method get_variant_remote()

Get and store the (remote) variant data.

Usage

Variant$get_variant_remote()

Method new()

Initialize this variant.

Usage

Variant$new(connect, content, key)

Arguments

connect

The Connect instance.

content

The Content instance.

key

The variant key.

Method send_mail()

Mail previously rendered content.

Usage

Variant$send_mail(to = c("me", "collaborators", "collaborators_viewers"))

Arguments

to

Targeting.

Method get_schedule()

Get the (remote) schedule data.

Usage

Variant$get_schedule()

Method get_schedule_remote()

Get the (remote) schedule data.

Usage

Variant$get_schedule_remote()

Method get_subscribers()

Get the subscribers.

Usage

Variant$get_subscribers()

Method remove_subscriber()

Remove a named subscriber.

Usage

Variant$remove_subscriber(guid)

Arguments

guid

User GUID.

Method add_subscribers()

Add named subscribers.

Usage

Variant$add_subscribers(guids)

Arguments

guids

User GUIDs.

Method render()

Render this variant.

Usage

Variant$render()

Method renderings()

List the renderings of this variant.

Usage

Variant$renderings()

Method update_variant()

Update this variant.

Usage

Variant$update_variant(...)

Arguments

...

Target fields and values.

Method jobs()

Jobs for this variant.

Usage

Variant$jobs()

Method job()

Return single job for this variant.

Usage

Variant$job(key)

Arguments

key

The job key.

Method get_url()

Return the URL for this variant.

Usage

Variant$get_url()

Method get_url_rev()

Return the URL associated with one rendering for this variant.

Usage

Variant$get_url_rev(rev)

Arguments

rev

Rendering identifier.

Method get_dashboard_url()

Return the URL for this variant in the Posit Connect dashboard.

Usage

Variant$get_dashboard_url(pane = "access")

Arguments

pane

The pane in the dashboard to link to.

Method print()

Print this object.

Usage

Variant$print(...)

Arguments

...

Unused.

Method clone()

The objects of this class are cloneable with this method.

Usage

Variant$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.